sharepoint/pyproject.toml
[tool.poetry]
name = "sharepoint"
version = "0.1.0"
description = "Connects Cohere to Sharepoint"
authors = ["Scott Mountenay <scott@lightsonsoftware.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "2.2.5"
azure-core = "1.23.1"
azure-identity = "1.10.0"
python-dotenv = "1.0.0"
msgraph-core = "0.2.2"
connexion = {version = "2.14.2", extras = ["swagger-ui"]}
requests = "^2.31.0"
gunicorn = "^21.2.0"
aiohttp = "^3.9.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
